==> notebooks/04_industrial_iot/0.0-nk-gdn-env-check.ipynb <==
Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.

SPDX-License-Identifier: Apache-2.0

# GDN Environment verification

The Graph Neural Network-Based Anomaly Detection (GDN) in Multivariate Time Series module is sourced from the [GDN](https://github.com/d-ailin/GDN) repository.

Installation of GDN requires `torch==1.5.0`, `torch-scatter`, `torch-sparse`, `torch-cluster`, `torch-spline-conv`, and `torch-geometric==1.5.0`.

The `anomaly-detection-spatial-temporal-data/prepare_gdn_environment.sh` script prepares the `kedro-gdn-venv` virtual environment and registers it with the Jupyter registry as a custom kernel.

### Check GDNTrainer import

import sys 
sys.path.append('../../src/anomaly_detection_spatial_temporal_data/')

from model.GDN.GDNTrainer import GDNTrainer

### Check driver functions import

from pipelines.iot_gdn_data_processing.nodes import clean_column_names, get_sensor_columns, save_sensor_cols
