==> notebooks/04_industrial_iot/0.0-nk-ncad-env-check.ipynb <==
Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.

SPDX-License-Identifier: Apache-2.0

# NCAD environment verification

The Neural Contextual Anomaly Detection (NCAD) for Time Series module is sourced from the [gluon-ts](https://github.com/Francois-Aubet/gluon-ts/tree/adding_ncad_to_nursery/src/gluonts/nursery/ncad) github repository.

The NCAD authors have configured a `setup.py` allowing NCAD to be installed with `pip install -e ncad`. Additionally, we need to install `torchmetrics==0.6.0` and `pytorch-lightning==1.3.8`.

The `anomaly-detection-spatial-temporal-data/prepare_ncad_environment.sh` script prepares the `kedro-ncad-venv` virtual environment and registers it with the Jupyter registry as a custom kernel.

### Check NCAD import

import ncad

### Check driver functions import

import sys 
sys.path.append('../../src/anomaly_detection_spatial_temporal_data/')

from pipelines.iot_gdn_data_processing.nodes import clean_column_names, get_sensor_columns, save_sensor_cols
