==> notebooks/04_industrial_iot/2.2-nk-model-train-and-inference-nab.ipynb <==
Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.

SPDX-License-Identifier: Apache-2.0

## Table of Contents

1. Specify data folder containing individual CSVs
2. Specify location containing label JSON
3. Train Context OSE Model from NAB library
4. Perform inference on test set

import sys
sys.path.append("../../src/")

from anomaly_detection_spatial_temporal_data.model.time_series import NABAnomalyDetector

import pandas as pd

# Train NAB Context OSE model

model_name = "contextOSE"
model_path = "../../src/anomaly_detection_spatial_temporal_data/model/NAB"
input_dir = "../../data/02_intermediate/iot/ts_data"
label_file = "../../data/02_intermediate/iot/ts_label/labels-test.json"
output_dir =  "../../data/07_model_output/iot/ts_result-notebook"

nab = NABAnomalyDetector(
    model_name, 
    model_path,
    input_dir,
    label_file,
    output_dir,
)

nab.predict()

# Load an inference result
The NAB time series anomaly models use the historical context in a specific time series to identify if a future time step is an anomaly. Each time series is treated independently, so we do not expect this model to perform well on this dataset.

!ls ../../data/07_model_output/iot/ts_result-notebook/contextOSE/

output_csv = "contextOSE_L_T1.csv"

df_out = pd.read_csv(f"{output_dir}/contextOSE/{output_csv}")

df_out.head(20)

# References
Riccardo Taormina and Stefano Galelli and Nils Ole Tippenhauer and Elad Salomons and Avi Ostfeld and Demetrios G. Eliades and Mohsen Aghashahi and Raanju Sundararajan and Mohsen Pourahmadi and M. Katherine Banks and B. M. Brentan and Enrique Campbell and G. Lima and D. Manzi and D. Ayala-Cabrera and M. Herrera and I. Montalvo and J. Izquierdo and E. Luvizotto and Sarin E. Chandy and Amin Rasekh and Zachary A. Barker and Bruce Campbell and M. Ehsan Shafiee and Marcio Giacomoni and Nikolaos Gatsis and Ahmad Taha and Ahmed A. Abokifa and Kelsey Haddad and Cynthia S. Lo and Pratim Biswas and M. Fayzul K. Pasha and Bijay Kc and Saravanakumar Lakshmanan Somasundaram and Mashor Housh and Ziv Ohar; "The Battle Of The Attack Detection Algorithms: Disclosing Cyber Attacks On Water Distribution Networks." Journal of Water Resources Planning and Management, 144 (8), August 2018

Alexander Lavin and Subutai Ahmad. 2015. Evaluating Real-Time Anomaly Detection Algorithms – The Numenta Anomaly Benchmark.
